stage2/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
web3>=6.20.0
fastapi>=0.115.0
uvicorn>=0.30.0
